+++ app/src/main/java/com/huelvadevelopers/proyectozero/model/BankAccount.kt
@@ -6,7 +6,7 @@ import java.io.FileDescriptor
  * Created by DrAP on 07/07/2017.
  */
 
-class BankAccount (val id: Int, var name: String, var description: String, var balance: Float,
+class BankAccount (val id: Int, var name: String, var description: String, var balance: Double,
                    val currency: String, val color : Int){
 
 }
